+++ partiql-ast-passes/src/name_resolver.rs
@@ -205,7 +205,7 @@ impl<'ast> Visitor<'ast> for NameResolver {
         let id = *self.current_node();
         self.enclosing_clause
             .entry(EnclosingClause::Query)
-            .or_insert_with(Vec::new)
+            .or_default()
             .push(id);
         self.enter_keyref();
         Traverse::Continue
@@ -273,24 +273,18 @@ impl<'ast> Visitor<'ast> for NameResolver {
         let id = *self.current_node();
         self.enclosing_clause
             .entry(EnclosingClause::FromLet)
-            .or_insert_with(Vec::new)
+            .or_default()
             .push(id);
         self.enter_keyref();
 
         // Scopes above this `FROM` in the AST are in-scope to use variables defined by this from
         for in_scope in self.id_path_to_root.iter().rev().skip(1) {
-            self.in_scope
-                .entry(*in_scope)
-                .or_insert_with(Vec::new)
-                .push(id);
+            self.in_scope.entry(*in_scope).or_default().push(id);
         }
 
         // This `FROM` item is in-scope of variables defined by any preceding items in this `FROM` (e.g., lateral joins)
         for in_scope in self.lateral_stack.last().unwrap() {
-            self.in_scope
-                .entry(id)
-                .or_insert_with(Vec::new)
-                .push(*in_scope);
+            self.in_scope.entry(id).or_default().push(*in_scope);
         }
 
         self.lateral_stack.last_mut().unwrap().push(id);
@@ -383,7 +377,49 @@ impl<'ast> Visitor<'ast> for NameResolver {
         Traverse::Continue
     }
 
+    fn enter_group_key(&mut self, _group_key: &'ast GroupKey) -> Traverse {
+        self.enter_keyref();
+        let id = *self.current_node();
+
+        if self
+            .enclosing_clause
+            .get(&EnclosingClause::FromLet)
+            .is_none()
+        {
+            self.errors.push(AstTransformError::IllegalState(
+                "group_key expects a FromLet enclosing clause".to_string(),
+            ))
+        }
+
+        self.enclosing_clause
+            .get(&EnclosingClause::FromLet)
+            .expect("EnclosingClause::FromLet")
+            .iter()
+            .for_each(|enclosing_clause| {
+                self.in_scope.entry(id).or_default().push(*enclosing_clause);
+            });
+
+        self.enclosing_clause
+            .entry(EnclosingClause::Query)
+            .or_default()
+            .push(id);
+        Traverse::Continue
+    }
+
     fn exit_group_key(&mut self, group_key: &'ast GroupKey) -> Traverse {
+        let KeyRefs {
+            consume,
+            produce_required,
+            ..
+        } = match self.exit_keyref() {
+            Ok(kr) => kr,
+            Err(e) => {
+                self.errors.push(e);
+                return Traverse::Stop;
+            }
+        };
+        let mut produce = produce_required;
+
         let id = *self.current_node();
         // get the "as" alias for each `GROUP BY` expr
         // 1. if explicitly given
@@ -397,21 +433,50 @@ impl<'ast> Visitor<'ast> for NameResolver {
             Symbol::Unknown(self.id_gen.next_id())
         };
         self.aliases.insert(id, as_alias.clone());
+        produce.insert(as_alias.clone());
         self.keyref_stack
             .last_mut()
             .unwrap()
             .produce_required
             .insert(as_alias);
+        self.schema.insert(id, KeySchema { consume, produce });
+        Traverse::Continue
+    }
+
+    fn enter_group_by_expr(&mut self, _group_by_expr: &'ast GroupByExpr) -> Traverse {
+        self.enter_keyref();
+        let id = *self.current_node();
+        // Scopes above this `GROUP BY` in the AST are in-scope to use variables defined by this GROUP BY
+        for in_scope in self.id_path_to_root.iter().rev().skip(1) {
+            self.in_scope.entry(*in_scope).or_default().push(id);
+        }
         Traverse::Continue
     }
 
     fn exit_group_by_expr(&mut self, group_by_expr: &'ast GroupByExpr) -> Traverse {
+        let id = *self.current_node();
+        let KeyRefs {
+            consume,
+            produce_required,
+            ..
+        } = match self.exit_keyref() {
+            Ok(kr) => kr,
+            Err(e) => {
+                self.errors.push(e);
+                return Traverse::Stop;
+            }
+        };
+
+        // TODO: delete in_scope for FROM sources in subsequent clauses
+
+        let mut produce: Names = produce_required;
         // add the `GROUP AS` alias
         if let Some(sym) = &group_by_expr.group_as_alias {
-            let id = *self.current_node();
             let as_alias = Symbol::Known(sym.clone());
-            self.aliases.insert(id, as_alias);
+            self.aliases.insert(id, as_alias.clone());
+            produce.insert(as_alias);
         }
+        self.schema.insert(id, KeySchema { consume, produce });
         Traverse::Continue
     }
 }