+++ partiql-eval/src/lib.rs
@@ -18,7 +18,7 @@ mod tests {
     use crate::plan::EvaluationMode;
     use partiql_logical::{
         BagExpr, BetweenExpr, BinaryOp, BindingsOp, CoalesceExpr, ExprQuery, IsTypeExpr, JoinKind,
-        ListExpr, LogicalPlan, NullIfExpr, PathComponent, TupleExpr, Type, ValueExpr,
+        ListExpr, LogicalPlan, NullIfExpr, PathComponent, TupleExpr, Type, ValueExpr, VarRefType,
     };
     use partiql_value as value;
     use partiql_value::Value::{Missing, Null};
@@ -68,7 +68,10 @@ mod tests {
 
     fn scan(name: &str, as_key: &str) -> BindingsOp {
         BindingsOp::Scan(logical::Scan {
-            expr: ValueExpr::VarRef(BindingsName::CaseInsensitive(name.into())),
+            expr: ValueExpr::VarRef(
+                BindingsName::CaseInsensitive(name.into()),
+                VarRefType::Global,
+            ),
             as_key: as_key.to_string(),
             at_key: None,
         })
@@ -76,9 +79,10 @@ mod tests {
 
     fn path_var(name: &str, component: &str) -> ValueExpr {
         ValueExpr::Path(
-            Box::new(ValueExpr::VarRef(BindingsName::CaseInsensitive(
-                name.into(),
-            ))),
+            Box::new(ValueExpr::VarRef(
+                BindingsName::CaseInsensitive(name.into()),
+                VarRefType::Local,
+            )),
             vec![PathComponent::Key(BindingsName::CaseInsensitive(
                 component.to_string(),
             ))],
@@ -153,7 +157,10 @@ mod tests {
     fn eval_bin_op(op: BinaryOp, lhs: Value, rhs: Value, expected_first_elem: Value) {
         let mut plan = LogicalPlan::new();
         let scan = plan.add_operator(BindingsOp::Scan(logical::Scan {
-            expr: ValueExpr::VarRef(BindingsName::CaseInsensitive("data".into())),
+            expr: ValueExpr::VarRef(
+                BindingsName::CaseInsensitive("data".into()),
+                VarRefType::Global,
+            ),
             as_key: "data".to_string(),
             at_key: None,
         }));
@@ -164,9 +171,10 @@ mod tests {
                 ValueExpr::BinaryExpr(
                     op,
                     Box::new(ValueExpr::Path(
-                        Box::new(ValueExpr::VarRef(BindingsName::CaseInsensitive(
-                            "data".into(),
-                        ))),
+                        Box::new(ValueExpr::VarRef(
+                            BindingsName::CaseInsensitive("data".into()),
+                            VarRefType::Local,
+                        )),
                         vec![PathComponent::Key(BindingsName::CaseInsensitive(
                             "lhs".to_string(),
                         ))],
@@ -660,7 +668,10 @@ mod tests {
         fn eval_between_op(value: Value, from: Value, to: Value, expected_first_elem: Value) {
             let mut plan = LogicalPlan::new();
             let scan = plan.add_operator(BindingsOp::Scan(logical::Scan {
-                expr: ValueExpr::VarRef(BindingsName::CaseInsensitive("data".into())),
+                expr: ValueExpr::VarRef(
+                    BindingsName::CaseInsensitive("data".into()),
+                    VarRefType::Global,
+                ),
                 as_key: "data".to_string(),
                 at_key: None,
             }));
@@ -670,9 +681,10 @@ mod tests {
                     "result".to_string(),
                     ValueExpr::BetweenExpr(BetweenExpr {
                         value: Box::new(ValueExpr::Path(
-                            Box::new(ValueExpr::VarRef(BindingsName::CaseInsensitive(
-                                "data".into(),
-                            ))),
+                            Box::new(ValueExpr::VarRef(
+                                BindingsName::CaseInsensitive("data".into()),
+                                VarRefType::Local,
+                            )),
                             vec![PathComponent::Key(BindingsName::CaseInsensitive(
                                 "value".to_string(),
                             ))],
@@ -877,7 +889,7 @@ mod tests {
         let project = lg.add_operator(Project(logical::Project {
             exprs: Vec::from([(
                 "r".to_string(),
-                ValueExpr::VarRef(BindingsName::CaseInsensitive("r".into())),
+                ValueExpr::VarRef(BindingsName::CaseInsensitive("r".into()), VarRefType::Local),
             )]),
         }));
 
@@ -1125,7 +1137,10 @@ mod tests {
     fn eval_is_op(not: bool, expr: Value, is_type: Type, expected_first_elem: Value) {
         let mut plan = LogicalPlan::new();
         let scan = plan.add_operator(BindingsOp::Scan(logical::Scan {
-            expr: ValueExpr::VarRef(BindingsName::CaseInsensitive("data".into())),
+            expr: ValueExpr::VarRef(
+                BindingsName::CaseInsensitive("data".into()),
+                VarRefType::Global,
+            ),
             as_key: "data".to_string(),
             at_key: None,
         }));
@@ -1136,9 +1151,10 @@ mod tests {
                 ValueExpr::IsTypeExpr(IsTypeExpr {
                     not,
                     expr: Box::new(ValueExpr::Path(
-                        Box::new(ValueExpr::VarRef(BindingsName::CaseInsensitive(
-                            "data".into(),
-                        ))),
+                        Box::new(ValueExpr::VarRef(
+                            BindingsName::CaseInsensitive("data".into()),
+                            VarRefType::Local,
+                        )),
                         vec![PathComponent::Key(BindingsName::CaseInsensitive(
                             "expr".to_string(),
                         ))],
@@ -1188,7 +1204,10 @@ mod tests {
     fn eval_null_if_op(lhs: Value, rhs: Value, expected_first_elem: Value) {
         let mut plan = LogicalPlan::new();
         let scan = plan.add_operator(BindingsOp::Scan(logical::Scan {
-            expr: ValueExpr::VarRef(BindingsName::CaseInsensitive("data".into())),
+            expr: ValueExpr::VarRef(
+                BindingsName::CaseInsensitive("data".into()),
+                VarRefType::Global,
+            ),
             as_key: "data".to_string(),
             at_key: None,
         }));
@@ -1198,9 +1217,10 @@ mod tests {
                 "result".to_string(),
                 ValueExpr::NullIfExpr(NullIfExpr {
                     lhs: Box::new(ValueExpr::Path(
-                        Box::new(ValueExpr::VarRef(BindingsName::CaseInsensitive(
-                            "data".into(),
-                        ))),
+                        Box::new(ValueExpr::VarRef(
+                            BindingsName::CaseInsensitive("data".into()),
+                            VarRefType::Local,
+                        )),
                         vec![PathComponent::Key(BindingsName::CaseInsensitive(
                             "lhs".to_string(),
                         ))],
@@ -1243,16 +1263,20 @@ mod tests {
     fn eval_coalesce_op(elements: Vec<Value>, expected_first_elem: Value) {
         let mut plan = LogicalPlan::new();
         let scan = plan.add_operator(BindingsOp::Scan(logical::Scan {
-            expr: ValueExpr::VarRef(BindingsName::CaseInsensitive("data".into())),
+            expr: ValueExpr::VarRef(
+                BindingsName::CaseInsensitive("data".into()),
+                VarRefType::Global,
+            ),
             as_key: "data".to_string(),
             at_key: None,
         }));
 
         fn index_to_valueexpr(i: usize) -> ValueExpr {
             ValueExpr::Path(
-                Box::new(ValueExpr::VarRef(BindingsName::CaseInsensitive(
-                    "data".into(),
-                ))),
+                Box::new(ValueExpr::VarRef(
+                    BindingsName::CaseInsensitive("data".into()),
+                    VarRefType::Local,
+                )),
                 vec![PathComponent::Key(BindingsName::CaseInsensitive(format!(
                     "arg{i}"
                 )))],
@@ -1394,9 +1418,10 @@ mod tests {
             exprs: Vec::from([(
                 "b".to_string(),
                 ValueExpr::Path(
-                    Box::new(ValueExpr::VarRef(BindingsName::CaseInsensitive(
-                        "data".into(),
-                    ))),
+                    Box::new(ValueExpr::VarRef(
+                        BindingsName::CaseInsensitive("data".into()),
+                        VarRefType::Local,
+                    )),
                     vec![PathComponent::Key(BindingsName::CaseInsensitive(
                         "a".to_string(),
                     ))],
@@ -1922,9 +1947,10 @@ mod tests {
             expr: ValueExpr::BinaryExpr(
                 BinaryOp::Gt,
                 Box::new(ValueExpr::Path(
-                    Box::new(ValueExpr::VarRef(BindingsName::CaseInsensitive(
-                        "customer".into(),
-                    ))),
+                    Box::new(ValueExpr::VarRef(
+                        BindingsName::CaseInsensitive("customer".into()),
+                        VarRefType::Local,
+                    )),
                     vec![PathComponent::Key(BindingsName::CaseInsensitive(
                         "balance".to_string(),
                     ))],
@@ -1938,9 +1964,10 @@ mod tests {
                 (
                     "firstName".to_string(),
                     ValueExpr::Path(
-                        Box::new(ValueExpr::VarRef(BindingsName::CaseInsensitive(
-                            "customer".into(),
-                        ))),
+                        Box::new(ValueExpr::VarRef(
+                            BindingsName::CaseInsensitive("customer".into()),
+                            VarRefType::Local,
+                        )),
                         vec![PathComponent::Key(BindingsName::CaseInsensitive(
                             "firstName".to_string(),
                         ))],
@@ -1951,17 +1978,19 @@ mod tests {
                     ValueExpr::BinaryExpr(
                         BinaryOp::Concat,
                         Box::new(ValueExpr::Path(
-                            Box::new(ValueExpr::VarRef(BindingsName::CaseInsensitive(
-                                "customer".into(),
-                            ))),
+                            Box::new(ValueExpr::VarRef(
+                                BindingsName::CaseInsensitive("customer".into()),
+                                VarRefType::Local,
+                            )),
                             vec![PathComponent::Key(BindingsName::CaseInsensitive(
                                 "firstName".to_string(),
                             ))],
                         )),
                         Box::new(ValueExpr::Path(
-                            Box::new(ValueExpr::VarRef(BindingsName::CaseInsensitive(
-                                "customer".into(),
-                            ))),
+                            Box::new(ValueExpr::VarRef(
+                                BindingsName::CaseInsensitive("customer".into()),
+                                VarRefType::Local,
+                            )),
                             vec![PathComponent::Key(BindingsName::CaseInsensitive(
                                 "firstName".to_string(),
                             ))],
@@ -2003,9 +2032,10 @@ mod tests {
             expr: ValueExpr::BinaryExpr(
                 BinaryOp::In,
                 Box::new(ValueExpr::Path(
-                    Box::new(ValueExpr::VarRef(BindingsName::CaseInsensitive(
-                        "data".into(),
-                    ))),
+                    Box::new(ValueExpr::VarRef(
+                        BindingsName::CaseInsensitive("data".into()),
+                        VarRefType::Local,
+                    )),
                     vec![PathComponent::Key(BindingsName::CaseInsensitive(
                         "a".to_string(),
                     ))],
@@ -2018,9 +2048,10 @@ mod tests {
             exprs: Vec::from([(
                 "b".to_string(),
                 ValueExpr::Path(
-                    Box::new(ValueExpr::VarRef(BindingsName::CaseInsensitive(
-                        "data".into(),
-                    ))),
+                    Box::new(ValueExpr::VarRef(
+                        BindingsName::CaseInsensitive("data".into()),
+                        VarRefType::Local,
+                    )),
                     vec![PathComponent::Key(BindingsName::CaseInsensitive(
                         "a".to_string(),
                     ))],
@@ -2179,7 +2210,7 @@ mod tests {
 
     mod clause_from {
         use crate::eval::evaluable::{EvalScan, Evaluable};
-        use crate::eval::expr::{EvalPath, EvalPathComponent, EvalVarRef};
+        use crate::eval::expr::{EvalGlobalVarRef, EvalPath, EvalPathComponent};
         use crate::eval::BasicContext;
         use partiql_value::{bag, list, BindingsName};
 
@@ -2202,7 +2233,7 @@ mod tests {
             let ctx = BasicContext::new(p0);
 
             let mut scan = EvalScan::new_with_at_key(
-                Box::new(EvalVarRef {
+                Box::new(EvalGlobalVarRef {
                     name: BindingsName::CaseInsensitive("someOrderedTable".to_string()),
                 }),
                 "x",
@@ -2228,7 +2259,7 @@ mod tests {
             let ctx = BasicContext::new(p0);
 
             let mut scan = EvalScan::new_with_at_key(
-                Box::new(EvalVarRef {
+                Box::new(EvalGlobalVarRef {
                     name: BindingsName::CaseInsensitive("someUnorderedTable".to_string()),
                 }),
                 "x",
@@ -2257,7 +2288,7 @@ mod tests {
             let mut p0: MapBindings<Value> = MapBindings::default();
             p0.insert("someOrderedTable", some_ordered_table().into());
 
-            let table_ref = EvalVarRef {
+            let table_ref = EvalGlobalVarRef {
                 name: BindingsName::CaseInsensitive("someOrderedTable".to_string()),
             };
             let path_to_scalar = EvalPath {
@@ -2282,7 +2313,7 @@ mod tests {
             let mut p0: MapBindings<Value> = MapBindings::default();
             p0.insert("someOrderedTable", some_ordered_table().into());
 
-            let table_ref = EvalVarRef {
+            let table_ref = EvalGlobalVarRef {
                 name: BindingsName::CaseInsensitive("someOrderedTable".to_string()),
             };
             let path_to_scalar = EvalPath {
@@ -2306,7 +2337,7 @@ mod tests {
         use partiql_value::{bag, BindingsName, Tuple};
 
         use crate::eval::evaluable::{EvalUnpivot, Evaluable};
-        use crate::eval::expr::EvalVarRef;
+        use crate::eval::expr::EvalGlobalVarRef;
         use crate::eval::BasicContext;
 
         use super::*;
@@ -2322,7 +2353,7 @@ mod tests {
             p0.insert("justATuple", just_a_tuple().into());
 
             let mut unpivot = EvalUnpivot::new(
-                Box::new(EvalVarRef {
+                Box::new(EvalGlobalVarRef {
                     name: BindingsName::CaseInsensitive("justATuple".to_string()),
                 }),
                 "price",
@@ -2346,7 +2377,7 @@ mod tests {
             p0.insert("nonTuple", Value::from(1));
 
             let mut unpivot = EvalUnpivot::new(
-                Box::new(EvalVarRef {
+                Box::new(EvalGlobalVarRef {
                     name: BindingsName::CaseInsensitive("nonTuple".to_string()),
                 }),
                 "x",